Reorganize database layout
